Fix relacion device_id a BigInteger

+++ app/models/event.py
@@ -12,7 +12,7 @@ class Event(BaseModel):
     date = db.Column(db.Date())
     app_version_code = db.Column(db.String(10))
     sim_serial_number = db.Column(db.BigInteger, db.ForeignKey("sims.serial_number"))
-    device_id = db.Column(db.Integer, db.ForeignKey("devices.device_id"))
+    device_id = db.Column(db.BigInteger, db.ForeignKey("devices.device_id"))
 
     # Herencia
     type = db.Column(db.String(50))
